separate integration and unit tests, use codecov flags
