+++ .travis.yml
@@ -17,16 +17,19 @@ matrix:
       env: NO_WATCH_TESTS=1 JOB_PART=lint
     - os: linux
       node_js: "8"
-      env: NO_WATCH_TESTS=1 JOB_PART=test
+      env: NO_WATCH_TESTS=1 JOB_PART=integration
+    - os: linux
+      node_js: "8"
+      env: NO_WATCH_TESTS=1 JOB_PART=unit
     - os: linux
       node_js: "6"
-      env: NO_WATCH_TESTS=1 JOB_PART=test
+      env: NO_WATCH_TESTS=1 JOB_PART=integration
     - os: linux
       node_js: "4"
-      env: NO_WATCH_TESTS=1 JOB_PART=test
+      env: NO_WATCH_TESTS=1 JOB_PART=integration
     - os: osx
       node_js: "8"
-      env: NO_WATCH_TESTS=1 JOB_PART=test
+      env: NO_WATCH_TESTS=1 JOB_PART=integration
   allow_failures:
     - os: osx
   fast_finish: true
@@ -38,7 +41,7 @@ script: npm run travis:$JOB_PART
 
 after_success:
   - cat ./coverage/lcov.info | node_modules/.bin/coveralls --verbose
-  - cat ./coverage/coverage.json | node_modules/codecov.io/bin/codecov.io.js
+  - bash <(curl -s https://codecov.io/bash) -F $JOB_PART
   - rm -rf ./coverage
 
 notifications:
