+++ codecov.yml
@@ -6,12 +6,29 @@ coverage:
   range: "70...100"
   status:
     project:
-      default:
+      default: off
+      integration:
+        flags: integration
         target: auto
+      unit:
+        flags: unit
+        target: 0%
     patch:
-      default:
-        target: auto
+      default: off
+      integration:
+        flags: integration
+        target: 90%
+        base: pr
+      unit:
+        flags: unit
+        target: 0%
+        base: pr
     changes:
-      default:
-        branches: null
+      default: off
+      integration:
+        flags: integration
+        target: 0%
+      unit:
+        flags: unit
+        target: 0%
 comment: off
