+++ package.json
@@ -33,7 +33,6 @@
     "benchmark": "^2.1.1",
     "bundle-loader": "~0.5.0",
     "codacy-coverage": "^2.0.1",
-    "codecov.io": "^0.1.2",
     "coffee-loader": "~0.7.1",
     "coffee-script": "^1.10.0",
     "coveralls": "^2.11.2",
@@ -58,6 +57,7 @@
     "raw-loader": "~0.5.0",
     "react": "^15.2.1",
     "react-dom": "^15.2.1",
+    "rimraf": "^2.6.2",
     "script-loader": "~0.7.0",
     "should": "^11.1.1",
     "simple-git": "^1.65.0",
@@ -89,13 +89,17 @@
     "schemas/"
   ],
   "scripts": {
-    "test": "mocha test/*.test.js --max-old-space-size=4096 --harmony --check-leaks",
-    "travis:test": "npm run cover:min",
+    "test": "mocha test/*.test.js test/*.unittest.js --max-old-space-size=4096 --harmony --trace-deprecation --check-leaks",
+    "test:integration": "mocha test/*.test.js --max-old-space-size=4096 --harmony --trace-deprecation --check-leaks",
+    "test:unit": "mocha test/*.unittest.js --max-old-space-size=4096 --harmony --trace-deprecation --check-leaks",
+    "travis:integration": "npm run cover:init && npm run cover:integration && npm run cover:report-min",
+    "travis:unit": "npm run cover:init && npm run cover:unit && npm run cover:report-min",
     "travis:lint": "npm run lint-files && npm run nsp",
     "travis:benchmark": "npm run benchmark",
-    "appveyor:test": "node node_modules\\mocha\\bin\\mocha --max-old-space-size=4096 --harmony test/*.test.js",
+    "appveyor:integration": "npm run cover:init && npm run cover:integration && npm run cover:report-min",
+    "appveyor:unit": "npm run cover:init && npm run cover:unit && npm run cover:report-min",
     "appveyor:benchmark": "npm run benchmark",
-    "circleci:test": "node node_modules/mocha/bin/mocha --max-old-space-size=4096 --harmony test/*.test.js",
+    "circleci:test": "node node_modules/mocha/bin/mocha --max-old-space-size=4096 --harmony --trace-deprecation test/*.test.js test/*.unittest.js",
     "circleci:lint": "npm run lint-files && npm run nsp",
     "build:examples": "cd examples && node buildAll.js",
     "pretest": "npm run lint-files",
@@ -104,9 +108,14 @@
     "fix": "npm run lint -- --fix",
     "beautify-lint": "beautify-lint \"lib/**/*.js\" \"hot/**/*.js\" \"bin/**/*.js\" \"benchmark/*.js\" \"test/*.js\"",
     "nsp": "nsp check --output summary",
-    "benchmark": "mocha --max-old-space-size=4096 --harmony test/*.benchmark.js -R spec",
-    "cover": "node --max-old-space-size=4096 --harmony ./node_modules/istanbul/lib/cli.js cover -x '**/*.runtime.js' node_modules/mocha/bin/_mocha -- test/*.test.js",
-    "cover:min": "node --max-old-space-size=4096 --harmony ./node_modules/istanbul/lib/cli.js cover -x '**/*.runtime.js' --report lcovonly node_modules/mocha/bin/_mocha -- test/*.test.js",
+    "benchmark": "mocha --max-old-space-size=4096 --harmony --trace-deprecation test/*.benchmark.js -R spec",
+    "cover": "npm run cover:init && npm run cover:all && npm run cover:report",
+    "cover:init": "rimraf coverage",
+    "cover:all": "node --max-old-space-size=4096 --harmony --trace-deprecation ./node_modules/istanbul/lib/cli.js cover --report none node_modules/mocha/bin/_mocha -- test/*.test.js test/*.unittest.js",
+    "cover:integration": "node --max-old-space-size=4096 --harmony --trace-deprecation ./node_modules/istanbul/lib/cli.js cover --report none node_modules/mocha/bin/_mocha -- test/*.test.js",
+    "cover:unit": "node --max-old-space-size=4096 --harmony --trace-deprecation ./node_modules/istanbul/lib/cli.js cover --report none node_modules/mocha/bin/_mocha -- test/*.unittest.js",
+    "cover:report": "istanbul report",
+    "cover:report-min": "istanbul report --report lcovonly",
     "publish-patch": "npm run lint && npm run beautify-lint && mocha && npm version patch && git push && git push --tags && npm publish"
   }
 }
