+++ test/AmdMainTemplatePlugin.unittest.js
