+++ test/CachePlugin.unittest.js
