+++ test/CaseSensitiveModulesWarning.unittest.js
