+++ test/ContextReplacementPlugin.unittest.js
