+++ test/DependenciesBlockVariable.unittest.js
