+++ test/HarmonyModulesHelpers.unittest.js
