+++ test/HotUpdateChunkTemplate.unittest.js
