+++ test/JsonpHotUpdateChunkTemplatePlugin.unittest.js
