+++ test/LibraryTemplatePlugin.unittest.js
