+++ test/NodeHotUpdateChunkTemplatePlugin.unittest.js
