+++ test/RequireJsStuffPlugin.unittest.js
