+++ test/SourceMapDevToolModuleOptionsPlugin.unittest.js
