+++ test/WebWorkerChunkTemplatePlugin.unittest.js
