+++ test/WebWorkerHotUpdateChunkTemplatePlugin.unittest.js
