+++ test/WebWorkerMainTemplatePlugin.unittest.js
