+++ test/compareLocations.unittest.js
