+++ test/removeAndDo.unittest.js
